feature #21 : new advanced configuration setting `batch_download_use_representative_for_ext` (array of extension strings)
For a non image file, like a TIFF or a PDF, and for a multiple size (anything but the original) let
the user decide if she wants the original anyway or a derivative from the representative.

By default, TIFF files will use their representative (if available). The webmaster can empty the
array, or set it to other extensions, like:

$conf['batch_download_use_representative_for_ext'] = array('pdf', 'PDF', 'svg', 'SVG');